TensorFlow.OpGen.AttrVal

Description

Wrapping of TensorFlow attributes into Haskell entities.

Documentation

data AttrDef

data AttrCase f

Constructors

AttrBytes (f ByteString)

AttrInt64 (f Int64)

AttrFloat (f Float)

AttrBool (f Bool)

AttrType (f DataType)

AttrShape (f TensorShapeProto)

data AttrTemplate

Type-reified representation of TensorFlow AttrDef. Initially limited to just the types in Op descriptors.

Constructors

AttrSingle (AttrCase Template)

AttrList (AttrCase [])

AttrTensor UnusedTensor

data Template a

Specifies the optional default value and a set of allowed values for the given type.

attrDef :: OpDef'AttrDef -> AttrDef

attrOriginal :: Lens' AttrDef OpDef'AttrDef

attrTemplate :: Lens' AttrDef AttrTemplate

templateDefault :: Lens' (Template a) (Maybe a)

templateRestrictions :: Lens' (Template a) [a]