TensorFlow.OpGen.ParsedOp

Description

This module helps parse the proto OpDef into a Haskell type which is more descriptive of how the attributes and arguments will be used in the generated code.

Documentation

data ParsedOp

Constructors

ParsedOp

Fields

parsedOpName :: Name

parsedOpSummary :: Text

parsedOpDescription :: Text

parsedInputs :: [ParsedArg]

parsedOutputs :: [ParsedArg]

explicitInputAttrs :: [Attr AttrType]

Attributes that must be set explicitly when creating the op. Associated with the type of the attribute.

inferredTypeAttrs :: [Attr TypeParam]

Attributes that are type parameters.

inferredListSizeAttrs :: [Attr (NonEmpty Name)]

parsedOpIsMonadic :: Bool

Whether this op is stateful or takes a stateful input. Such ops should not be CSE'd and must be monadic in our API (i.e., return a Build action).

data Name

Constructors

Name

Fields

haskellName :: HaskellName

tfName :: TFName

newtype HaskellName

A name that's appropriate for a variable in a Haskell source file.

Constructors

HaskellName

Fields

unHaskellName :: Text

newtype TFName

A raw name as specified in the OpDef proto.

Constructors

TFName

Fields

unTFName :: Text

data Attr a

A named attribute, associated with some information about it.

Constructors

Attr

Fields

attrName :: Name

attrDescription :: Text

attrInfo :: a

data AttrType

The type of an attribute.

Constructors

AttrSingle AttrBaseType

AttrList AttrBaseType

data AttrBaseType

Constructors

AttrBytes

AttrInt64

AttrFloat

AttrBool

AttrType

AttrShape

AttrTensor

data TypeParam

Constructors

TypeParam

Fields

typeParamIsList :: Bool

typeParamRestrictions :: Maybe (NonEmpty DataType)

The list of allowed types (see: TensorFlow.Types.OneOf). If Nothing, then any type is acceptable.

data ParsedArg

An input or output argument (Tensor) for an op.

Constructors

ParsedArg

Fields

parsedArgName :: Name

parsedArgDescription :: Text

parsedArgCase :: ParsedArgCase

data ParsedArgCase

Constructors

SimpleArg

Fields

argType :: ArgType

argCaseKind :: ArgKind

ListArg

Fields

argLength :: Name

The attribute that specifies this list's length.

argType :: ArgType

argCaseKind :: ArgKind

MixedListArg
A heterogeneous list.

Fields

argTypeAttr :: Name

argCaseKind :: ArgKind

ResourceArg

data ArgType

The type of an argument.

Constructors

ArgTypeFixed DataType
A fixed type.

ArgTypeAttr Name
A type that depends on an attribute.

data ArgKind

Constructors

ArgTensorRef

ArgTensorValue

ArgTensorBuild

ArgSomeTensor Text

argKind :: ParsedArgCase -> Maybe ArgKind

parseOp :: OpDef -> ParsedOp

camelCase :: Text -> Text